+++ apps/web/.next/server/server-reference-manifest.js
@@ -1,1 +1,1 @@
-self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"NqDWV70lUvChvgEX1Diz44A9O7rk5JnOX5GMcMBtX/A=\"\n}"
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"pv7yLPiYTOzJ2oa2ICzbFiHUJ+WfQo1Z+C9o3vsyPko=\"\n}"